backend/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Telegram bot
aiogram>=3.15.0
aiohttp>=3.11.0

# Database (async)
sqlalchemy>=2.0.0
asyncpg>=0.30.0
alembic>=1.14.0

# Cache / FSM storage
redis>=5.2.0

# Config
pydantic>=2.10.0
pydantic-settings>=2.6.0

# HTTP client (bot -> API)
httpx>=0.28.0

# Date parsing
python-dateutil>=2.9.0

# S3 (MinIO)
boto3>=1.35.0

# Messaging / background jobs
aio-pika>=9.4.0
celery[redis]>=5.4.0
